Refactor Deref()
